Use fixture for API testing
